Updated openstack/openstack Project: openstack/nova  633ed09f23c90636403e1d8610a578a37494ee79
libvirt: fix missed test in migration

The method get_updated_guest_xml has a specific test to ensure all
dependent methods are well called with the well parameters.  This
commit adds an assertion to ensure update_perf_events_xml is well
called when we want to retrieve the domain xml of a guest for
migration.

Change-Id: Ifda45b824fbc819173498be606622e8e663b2626