Updated openstack/openstack Project: openstack/nova  afa90849097b6e073d7a7b23193d2315ad0d4578
Deprecate os-certificates

Coming out of the Nova midcycle we decided we really needed to
deprecate the os-certificates API and service. It is not used by any
of the rest of Nova. It's an API that requires a dedicated daemon to
be run. It's making Nova assume the responsibility for generating
secure certificates, even though it's not really managing entropy in
any sane way (which is why nova-cert is a service, so you could stick
these on all your computes so as to not run out of entropy).

Change-Id: I0ffa9200d8b4f52da4751e731aecffe09fd3d4be