Updated openstack/openstack Project: openstack/nova  e75956043b54f911f202fd1e876cdfabc33647bd
Make Aggregate metadata functions work with API db

Adds the ability to update aggregate metadata where the aggregate is
contained in the API db. Modifications will be made to the 'aggregate_metadata'
table in the API db if that is where the aggregate is located or the cell
database otherwise.

blueprint cells-aggregate-api-db

Change-Id: I456beb753298f4aabbed6d20bea7cbc537842e50