Updated openstack/openstack Project: openstack/trove  5f2d81da09755e33ca5a42b343487b592e0ee62e
Fix user & database pagination in guests

Make sure we serialized paginated objects and not the other way round.
Ensure the 'next marker' is a string (name).
Encapsulate the logic in a utility function and add tests.

Update only guests that have been using the pagination functions.

Tested with scenario tests on Cassandra and Postgresql.

Change-Id: I0898270f1afcbe95a431492b1195fb3dd33e9d88
Closes-Bug: 1592099