Updated openstack/openstack Project: openstack/heat  b2e5cf9a18e855aa00196b44ac150bafc650bd69
Use handle_update_cancel() to cancel nested stack updates

The code designed to deal with the case where an in-progress StackResource
update is cancelled by stopping the update of the nested stack has never
worked because it was erroneously trying to catch StopIteration exceptions
instead of the actual exception raised when a generator is closed,
GeneratorExit.

This change moves the handling code to the new handle_update_cancel()
mechanism instead. Using this mechanism also means that StackResources will
be cancelled immediately, rather than waiting for up to error_wait_time
before cancelling the nested stack update. The previous patch adds a grace
period for IN_PROGRESS resources within the nested stack to complete, so we
want to notify it straight away to begin that period (and to not start any
new work).

Since the grace period on the child stack means that the child is likely to
still be IN_PROGRESS at the time we want to roll back, handle rollback by
sending an update-cancel with rollback message to the child. Only if the
child is no longer IN_PROGRESS will we start an update with the previous
template. If it is still IN_PROGRESS then just wait for it to reach
ROLLBACK_COMPLETE.

Change-Id: Ide210f695446fe8de2057b3b74c276165a6f9f3f
Closes-Bug: #1591337
Closes-Bug: #1446252