Updated openstack/openstack Project: openstack/tripleo-heat-templates  1eeb7a806999a85a20dfc86dd465135a885da397
Move nova constraints from controller manifest to each service

Currently we are still creating all the pacemaker constraints for nova
in the main overcloud_controller_pacemaker.pp manifest file.

Let's move those to each role where they belong. Note that given
that a constraint depends on two separate pacemaker resources it is a
bit arbitrary in which file they end up being (the one of the first
resource or the second one).

Depends-On: I96a3a313d15fac820b020feae0568437c2cbade3

Change-Id: I4f15485b1f355b3b38fc6c16552f204aa8bba7bc