Updated openstack/openstack Project: openstack/tripleo-heat-templates  9aec3de5b842967cfa104ec5aebd359b71379c64
Convert service_name to underscore syntax

Currently we use hyphens, e.g cinder-api, but in overcloud.yaml
we have a lot of references to services (e.g for AllNodesConfig)
by underscore, e.g cinder_api.  To enable dynamic generation of
this data, we need the service name in underscore format.

Change-Id: Ief13dfe5d8d7691dfe2534ad5c39d7eacbcb6f70