Updated openstack/openstack Project: openstack/heat-templates  051822a1795ab2e4314194448c39fc06931c7f4a
Correct perms of /var/run/heat-config/heat-config

/var/run isn't persisted during image building, so relying on
os-apply-config preserving the permissions of the existing file won't
work because there isn't one.

Instead, 55-heat-config needs to check and fix the permissions of this
file on each run. Once it is set by 55-heat-config, os-apply-config
will respect the corrected permissions when overwriting.

Change-Id: Iecce01d31d66f217ccf4b34cc29f05465dc1886d
Closes-Bug: #1520347