Updated openstack/openstack Project: openstack/tempest  2422ea3326733703b96e6f3e6b5b4855608d9ca1
Remove base_extensions_client for the reability

The service clients for the volume service are different from
the other services and it is a little difficult to read and
debug because that consists of multiple layers for v1 and v2.
That could be useful for the code optimization,  but actually
the merit seems small and unreadable.
This patch removes these layers from the volume clients.

Partially implements blueprint consistent-service-method-names

Change-Id: Ib9a89d94da30592ee22a1d60bf419c4d988471c3