Updated openstack/openstack Project: openstack/tempest  f6d163bdc9815739bf7a41c3db5fc9339ee150d5
Remove base_availability_zone_client for the reability

The service clients for the volume service are different from
the other services and it is a little difficult to read and
debug because that consists of multiple layers for v1 and v2.
That could be useful for the code optimization,  but actually
the merit seems small and unreadable.
This patch removes these layers from the volume clients.

Partially implements blueprint consistent-service-method-names

Change-Id: Ia51e2580af4df088fde71520890cb11c3e27ca86