Updated openstack/openstack Project: openstack/swift  0850b24327898ea5da9a0642d97576b2fd38ba1c
Simple code and improve the readability of the code

In a function,when a variable is defined at the front,it is unnecessary
to define the variable again.

Modify the variable name to improve the readability of the code.
For example: container_nodes=container_info["nodes"] instead of
containers=container_info["nodes"]

Change-Id: I30b887fed08ff650076ce8ce33b1502bc1795b3f