Updated openstack/openstack Project: openstack/neutron-vpnaas  f6903524e0684b603bb18ade983b5eb0ce4b33d4
Fix API Tests

Currently, when we run api tests with command: "tox -e api",
tests are failing with errors.

This change first fixes errors, make api tests pass. Then refactor
the neutron-vpnaas api tests tree to fit the tempest plugin.

As a tempest plugin, the steps to run tests by hands are:

1. Setup a local working environment for running tempest
    tempest init ${your_tempest_dir}

2. Enter ${your_tempest_dir}
    cd ${your_tempest_dir}

3. Check neutron_vpnaas_tests exist in tempest plugins:
    tempest list-plugins
    +----------------------+------------------------------------------------------+
    |         Name         |                      EntryPoint                      |
    +----------------------+------------------------------------------------------+
    |    neutron_tests     |  neutron.tests.tempest.plugin:NeutronTempestPlugin   |
    | neutron_vpnaas_tests | neutron_vpnaas.tests.tempest.plugin:VPNTempestPlugin |
    +----------------------+------------------------------------------------------+

4. Run neutron_vpnaas tests:
    tempest run --regex "^neutron_vpnaas.tests.tempest.api\."

In the jenkins gate,  devstack-gate/devstack-vm-gate-wrap.sh will invoke tempest
with proper configurations, such as:
    DEVSTACK_GATE_TEMPEST=1
    DEVSTACK_GATE_TEMPEST_ALL_PLUGINS=1
    DEVSTACK_GATE_TEMPEST_REGEX="^neutron_vpnaas.tests.tempest.api\."
The actual raw command in gate running under the tempest code directory is:
    tox -eall-plugin -- "^neutron_vpnaas.tests.tempest.api\."

This is a part of activate the API job:
https://review.openstack.org/#/c/337011/

Change-Id: Icdc946bccb77b4c78d161a590de1aa15e0a57139