Updated openstack/openstack Project: openstack/tempest  3570c1e90face7562b336b4437dd59b3a5d3f675
Skip test_reassign_port_between_servers until fixed

Since the neutron gate jobs were switched to xenial on
7/28 we have between 25% and 50% failure rates:

http://tinyurl.com/htuz8b2

There is a race where the interface detach from the first
guest is slow enough that the port.device_id goes to ''
(which is what the test is waiting for) before it attaches
the port to the second guest.

We're working a fix in nova to poll the guest after the
device detach (which is async in libvirt/qemu) until it's
gone at which point we'll then update the port's device_id
to be '', so everything will be delayed until the guest
actually reports the interface is detached.

Change-Id: I12553aa90e08e7d29aa37cbe253061b0d48752f8
Related-Bug: #1607714