Updated openstack/openstack Project: openstack/neutron  4fe49de5af032cb9802884bcb6073bd81b7eab78
Refactoring config options for common config opts

Refactoring neutron common config opts to be in neutron/conf/common so
that all the configuration options reside in a centralized location.
This simplifies the process of looking up the config opts and provides
an easy way to import.

Moved conf/common/config.py to conf/common.py as per review comments.

Partial-Bug: #1563069

Change-Id: Ib5fa294906549237630f87b9c848eebe0644088c