Updated openstack/openstack Project: openstack/keystonemiddleware  b735ce37a8f03d0fa11820c4d7342482aa1706a9
Remove oslo-incubator

Oslo-incubator has been deprecated. The only thing that we use from it
is the memorycache which is hopefully going away soon. Copy this
memorycache code into the _cache module so that we can refactor it as
necessary without worrying about oslo-incubator.

Involves some minor cleanups for pep8 fixes and making functions
private.

Change-Id: I7a19d4ded8b538b6ea02e4a08068c863705194a3