Updated openstack/openstack Project: openstack-infra/project-config  3a3935d5ec731119307ee2a9e57ef0627ef1de46
clean up release tool output

Now that the release tools are running in CI, some of the echo
statements we had in place for debugging during a manual run are
redundant with the "bash -x" mode. This change also removes the special
"title" function and replaces some of those calls with simpler echo
calls.

Change-Id: Ib70293f78078d91ebe0cc3f693d7bdefb7659041
Signed-off-by: Doug Hellmann <doug@doughellmann.com>