Updated openstack/openstack Project: openstack/oslotest  425d4651221d09141afd1de1367cf1ec188594ca
Import mock so that it works on Python 3.x

Running oslotest's tests fail on Python 3.x because of the "import mock"
lines in test_output.py and test_timeout.py.  That works for 2.x because
"mock" is an independent package, but in 3.x, it's a part of unittest in
the standard library.  This difference is covered in six.moves, which is
where mock is imported from in other test modules.

Change-Id: I3882c2a9be8abc93cf95942579cb9562c6377c01
Closes-Bug: #1607963