Updated openstack/openstack Project: openstack/nova  ae3d464f4ef36831ef7112bc58d466f9f0efa4c6
numa: remove the redundant check for hw_cpu/hw_mem list

we use a hw_numa_mem/cpu_set to check the hw_numa_mem/cpu list is
empty or not, actually, we only need to use 'if' to check the list
if empty. don't need the redundant flag. so remove it.

Change-Id: If81c842a05eecd5257c59c9324c3341be803e3d8