Updated openstack/openstack Project: openstack/trove  67940e90a68250f712e2dc753414b5ed2c7bfd9a
Change service candidates list for MariaDB

For MariaDB 10.1 on CentOS/RHEL, it is using mariadb.service
for systemd now instead of mysql.service. This change takes
this into account in how the service candidates list is managed.

Due to the change to service candidates, the restore and
replication strategies needed to change. Basically, now that
MariaDB is a full manager, it needs to have its "App" class
set properly (this is the restore strategy fix). Replication
broke because the replication strategy was using the Mysql
manager backup/restore strategy. This was fixed by doing a small
re-factor to add properties to the strategy base class to allow
for manager specific overrides.

Change-Id: I6aed32ebc704174aefe7543699169097532a1b55
Closes-bug: 1579110