Updated openstack/openstack Project: openstack/cinder  42f416462da077baa1d1729d98b6df23a17bdaa9
Add functional tests for nested quotas

Adds functional tests for nested quotas as an alternative
to creating a Tempest test + gate job that would only run
Cinder related quota checks.

NOTE: The QUOTA engine in Cinder is a global variable that lazy loads
the quota driver, so even if we change the config for the quota driver,
we won't reliably change the driver being used (or change it back)
unless the global variables get cleaned up, so using mock instead of
overriding _get_flags to change the quota driver.

Co-Authored-By: Xinli Guan <xinli@us.ibm.com>

Change-Id: Ic9b7ef3c8e4978c1f1fee38a5c2511ba5cbe5559