Updated openstack/openstack Project: openstack/neutron  7bce210f7891df8b6ef08bc5dd2f1cd8ae0de42b
Remove 'released-neutronclient' tag from official bug tags

It is usual that a bug is reported to released versions
of neutronclient and this tag does not make sense in general.

released-neutronclient was originally used to indicate that
bugs tagged with this tag have been fixed in a released version
though they are in 'Fix Committed' status. This tag was introduced
because launchpad sometimes times out when we tried to mark bugs
as 'Fix Released' and there is no way to distinguish them from
bugs fixed in an unreleased neutronclient.

Note that there is no open bugs tagged with the tag now,
so we can remove this tag safely.

Change-Id: Ibd0452a57bca1fee79a5fb8e52d116262cd0f256