Updated openstack/openstack Project: openstack-infra/project-config  5f299ef1ec499d26726c5291cf73ba1067ccce4f
Cross test constraints and project unittests.

Chnages to to upper-constraints.txt are run against a dsvm -tempest but
we've shown that some breakages pass tempest but fail in a projects
unittests.

This review adds the machinery and a proof-of-concept to run the py27
unittests in several projects against any review that alters
upper-constraints.txt.

There will always be a balance between covering projects and not
requiring vast resources in the gate.  This chnage selects a set of
projects that covers >50% of the tracked requirements.  We have a very
long-tail :(

Co-Authored-By: Clark Boylan <clark.boylan@gmail.com>
Change-Id: Id51e5d2e0d4f925c18b0390d6bfbf97b44f7b2c1