Updated openstack/openstack Project: openstack/sahara  ef15df0e2f3fc8ac8baa9e61567718979f8dadb8
improve error message for execution with retries

if execution failed because of exceeded number of retries
we will receive a little bit strange error like
'Operation create wasn't executed correctly after 0 attempts'.
That happens because we are using the same variable for iterating
in while loop and for error message creating.

Change-Id: Ie6e150863e11747bbfe30be5ab16401f878a5198