Updated openstack/openstack Project: openstack/glance_store  016351b36778892750661ebb08acd749c7072474
Split functional tests apart

Currently all the functional tests are ran together in one job. This
patch breaks the tests apart so only certain parts of them have to be ran.

This will mean the tests can be ran in parallel and only the particular
ones that are required, as the functional tests require that backend to
be avalible.

Change-Id: I61ba679b8041a89fb5a7589a62db8b6f42cd1058