Updated openstack/openstack Project: openstack/tripleo-heat-templates  071da7e880ef940c17fbd90579b48ed702e83876
Enable keystone to use the SSL middleware

The http_proxy_to_wsgi middleware was recently added to keystone as
default in the pipeline [1]. So this takes it into use instead of the
non-standard option we were using before, which will be deprecated.

We already enable this middleware for nova, cinder and heat.

[1] Iad628a863e55cbf20c89ef23ebc7527ba8e1a835

Depends-On: I0fec98a6e1d9c8be4d8b8df382b78ba2815790f9
Change-Id: I8c1b84adc828a2b8c9ea11c4e2b8349427b1b206