Updated openstack/openstack Project: openstack/tempest  11d8cd0fbc0aafc4fc410887f8f3e55c799e2a35
Set timeout value in urllib3.poolmanager.PoolManager

If timeout is not set in urllib3.poolmanager.PoolManager, it will
use "Python's default timeout for sockets", but if "timeout for
sockets" is not set, the timeout will be infinite(will not timeout).
so this is intented to set timeout value in urllib3.poolmanager.
PoolManager to avoid infinite timeout.

Change-Id: Ic035fdb93734c926b26b33feb610e0977e48c646
Closes-Bug: #1558931