Updated openstack/openstack Project: openstack/tempest  ef1a2428302159344100fa40de7d50798a744d40
Remove base_hosts_client for the reability

The service clients for the volume service are different from
the other services and it is a little difficult to read and
debug because that consists of multiple layers for v1 and v2.
That could be useful for the code optimization,  but actually
the merit seems small and unreadable.
This patch removes these layers from the volume clients.

Partially implements blueprint consistent-service-method-names

Change-Id: I176ca1fefcb93b1cc071ab02a53f89571d129b3f