Updated openstack/openstack Project: openstack/tempest  3cde1c6471af3cb8c41410a8b17300b00e21b34b
Remove base_services_client for the reability

The service clients for the volume service are different from
the other services and it is a little difficult to read and
debug because that consists of multiple layers for v1 and v2.
That could be useful for the code optimization,  but actually
the merit seems small and unreadable.
This patch removes these layers from the volume clients.

Partially implements blueprint consistent-service-method-names

Change-Id: I06661fa4b0eb9ed961a4fd2cc9b89462ef45c271