Updated openstack/openstack Project: openstack-infra/project-config  5fa5b350cba1fc302d674a8eeb1276ac83c7362e
zuul/layout/puppet: switch ovb to multinode for some modules

For some modules, we don't need to test baremetal deployment.
Actually, only Nova and Ironic do. So let's keep ovb job for them.

Though for the others, use nonha-multinode-nv which is faster and
will allow us later to use it in gate maybe (voting).

Also add the jobs to puppet-mistral and puppet-zaqar, that are deployed
by TripleO.

Change-Id: I4413a9054e0b2aacfec1097588461ff1e7b3f4fb