Updated openstack/openstack Project: openstack/manila  4c2a69c7e1737f0d4e79da900de41d7a95c21551
Fix share migration test with snapshot support

Test 'test_migrate_share_with_snapshot_v2_5' validates that
share migration returns an error when a share has a snapshot
created, but to do so, it tries to create a snapshot, which
would fail in a backend that does not support snapshot.
This fix adds a validation that skips the test in this situation.

Change-Id: Icd6b04266ac2de83747b99295cb3538967922d65
Closes-bug: #1608293