Updated openstack/openstack Project: openstack-infra/project-config  10b56f6d4872aa062024e5f22a0b2b510c30fda9
zuul/tripleo: move undercloud job to check pipeline

We managed to make the undercloud job working in experimental pipeline.
This patch aims to move it to the check pipeline.

It is not clear if we will keep this job in the future, as we are
working on a undercloud upgrade job. Though until we get there, this is
a first iteration of having a job that only deploys an undercloud.

Once we stabilize it, we will enable voting on it.

Change-Id: Ifcaaa7acaa9dedaf90b4cb0e604aeb7ad7bd0b34