Updated openstack/openstack Project: openstack/keystone  3e0c43ea4ef0b98aae1a51ba598807450f009b05
PCI-DSS Password history requirements

This patch satisfies the following PCI-DSS password history
requirements:

* PCI-DSS 8.2.5: Do not allow an individual to submit a new
password/phrase that is the same as any of the last four
passwords/phrases he or she has used.

This feature is configurable; thus, not allowing users to change their
password if it matches a configurable number of previous passwords.

Partially-implements: blueprint pci-dss
Change-Id: I3765cd71463de7f0c7b2c1066ca8848f194c94f1