Updated openstack/openstack Project: openstack/nova  4b774a52ddc0458dffc2dd1f794bfc0c7d9981fc
Use ovo in test_obj_make_compatible()

o.vo has a fixture for testing compatibility routines. This changes the
nova unit tests to use that compatibility testing on all objects. In the
original nova test, the versions tested were from 1.0 to 1.(x-1), where
1.x is the current version. o.vo's fixture tests from 1.0 to 1.x, which
should be a no-op, but is also a good thing to test (if we can't convert
1.x to 1.x, that's a bad thing).

o.vo also had an enhancement where it passed args and kwargs to __init__
in the different objects, which was a bug in nova, but using that
fixture allows us to also close the nova bug. This change in o.vo was
I62646b99adca47a9c9fe0f466f7a23ac8fa4553e.

Change-Id: Ie127f55f2e5bcd383796ff51253cf1091be16c98
Closes-Bug: #1537882