Updated openstack/openstack Project: openstack-infra/tripleo-ci  d8b33a7f4253ea122902ff60269642b62e90eebb
Don't assume an eth1

Not all cloud providers (osic and ovh in particular) spawn instances
with an eth0 and eth1, so don't assume an eth1 exists when finding it's
ip address to add to $no_proxy.

This recently started failing when we merged:
I2b84b3dfae60bbe7d6a76259b1d718b7dda08739

which added set -o pipefail to toci_instack_osinfra.sh and passed CI
when it ran on a different provider (rax).

Change-Id: I601a5e09eafbc1f03d7913ca6ffc060e3d774a44
Closes-Bug: #1609357