Updated openstack/openstack Project: openstack/nova  4a7c75d9f7ab97948efe79d1e9da042e90dbbdcd
Add server name verification in instance search

There are Nova API calls that can accept a server name
as a regex parameter. For example, if the command
'nova list --name <some_pattern>' is invoked with
an incorrect value of the pattern, the nova-api makes
a request to database and returns Http 500 error. Seems
it's not convenient.

This fix checks the pattern. If it isn't correct,
the nova-api returns error 400 with a prompt of a bad regex.

Change-Id: Iad63c668d09ec6a82ace29700fb4949c1acfbe1c
Closes-Bug: #1603728